requirements.txt
playwright>=1.40.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
